+++ app/jobs/email_debate_outcomes_job.rb
@@ -1,9 +1,4 @@
 class EmailDebateOutcomesJob < EmailPetitionSignatoriesJob
-  def self.run_later_tonight(petition)
-    petition.set_email_requested_at_for('debate_outcome', to: Time.current)
-    super(petition, petition.get_email_requested_at_for('debate_outcome'))
-  end
-
   def email_delivery_job_class
     EmailDeliveryJobs::DebateOutcome
   end
